build(node): move series helpers to a named file
Also transform `index.js` into a barrel
